HIVE-29035: reduce HMSCachingCatalog to a pure cache
Roll back the AccessLevel-based authorization recently added to
HMSCachingCatalog: remove the authz fields, methods, and per-operation
guards, delete HMSPrivilegeHelper and RangerPrivilegeHelper, and drop the
3-arg constructor. L1/L2 caching, the JMX MBean, and dropped-table ->
NoSuchTableException are unchanged.

Per-operation authorization belongs in IcebergAuthorizer, which already does
it right for stage-create; extending it to the other operations is deferred
to a follow-up PR. Until then, writes and cache-miss reads are authorized by
HMS and stage-create by IcebergAuthorizer; only cache-hit reads are unchecked
at the catalog level, which the follow-up closes.

Tests: replaced TestHMSCachingCatalogAuthz with TestHMSCachingCatalogCache
(pure-cache cases, 2-arg constructor).
