--- standalone-metastore/metastore-rest-catalog/src/main/java/org/apache/iceberg/rest/HMSPrivilegeHelper.java
@@ -1,62 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.apache.iceberg.rest;
-
-/**
- * Interface for a helper that determines the access level a user has on a table or namespace.
- */
-public interface HMSPrivilegeHelper {
-
-  /**
-   * The level of access a user has on a table.
-   */
-  enum AccessLevel {
-    NONE,
-    READ_ONLY,
-    READ_WRITE
-  }
-
-  /**
-   * Whether the helper was successfully initialized with an authorizer.
-   */
-  default boolean isAvailable() {
-    return false;
-  }
-
-  /**
-   * Determines the access level a user has on a given table by calling
-   * the Ranger showPrivileges API directly.
-   *
-   * @param dbName    the database name
-   * @param tableName the table name
-   * @param userName  the user name
-   * @return the access level (NONE, READ_ONLY, or READ_WRITE)
-   */
-   AccessLevel getAccessLevel(String dbName, String tableName, String userName);
-
-  /**
-   * Determines the access level a user has on a namespace (database).
-   *
-   * @param dbName   the database name
-   * @param userName the user name
-   * @return the access level (NONE, READ_ONLY, or READ_WRITE)
-   */
-   AccessLevel getNamespaceAccessLevel(String dbName, String userName);
-}
